# apps/api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Sales CRM API
  version: 0.1.0
  description: Contract for the multi-tenant Sales CRM. Product-facing brand text is supplied by runtime product configuration.
servers:
  - url: /api/v1
security:
  - cookieSession: []
paths:
...
  /auth/login:
    post:
      operationId: identityLogin
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: Authenticated session and CSRF token cookies have been set
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionView" }
        "202":
          description: Password accepted; MFA challenge must be completed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MFAChallenge" }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /auth/session:
    get:
      operationId: identityProbeSession
      security: []
      responses:
        "200":
          description: Same-origin browser session probe; unauthenticated is a normal result
          headers:
            Cache-Control:
              schema: { type: string, const: no-store }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionProbe" }
        "500": { $ref: "#/components/responses/Problem" }
  /auth/register:
    post:
      operationId: identityRegisterDevelopmentUser
      security: []
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    {
                      $ref: "#/components/schemas/DevelopmentRegistrationRequest",
                    },
                },
            },
        }
      responses:
        "201":
          {
            description: Development user registered,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/RegisteredUser" } },
              },
          }
        "404": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /auth/mfa/verify:
    post:
      operationId: identityVerifyMFALogin
      security: []
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/MFALoginRequest" } },
            },
        }
      responses:
        "200":
          {
            description: MFA verified and authenticated session set,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/SessionView" } },
              },
          }
        "401": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /auth/password-reset/request:
    post:
      operationId: identityRequestPasswordReset
      security: []
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/PasswordResetRequest" },
                },
            },
        }
      responses:
        "202":
          { description: Request accepted without revealing account existence }
        "429": { $ref: "#/components/responses/Problem" }
  /auth/password-reset/confirm:
    post:
      operationId: identityConfirmPasswordReset
      security: []
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/PasswordResetConfirmation" },
                },
            },
        }
      responses:
        "204": { description: Password reset and existing sessions revoked }
        "422": { $ref: "#/components/responses/Problem" }
        "429": { $ref: "#/components/responses/Problem" }
  /auth/logout:
    post:
      operationId: identityLogout
      responses:
        "204": { description: Session revoked }
        "401": { $ref: "#/components/responses/Problem" }
  /me:
    get:
      operationId: identityMe
      responses:
        "200":
          description: Current user and accessible workspaces
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SessionView" }
        "401": { $ref: "#/components/responses/Problem" }
    patch:
      operationId: identityUpdatePreferences
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserPreferences" }
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/Problem" }
  /me/password:
    put:
      operationId: identityChangePassword
      parameters:
        - { $ref: "#/components/parameters/CSRFToken" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/PasswordChangeRequest" },
                },
            },
        }
      responses:
        "204": { description: Password changed and all sessions revoked }
        "401": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /me/sessions:
    delete:
      operationId: identityLogoutAllSessions
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      responses:
        "204": { description: All user sessions revoked }
  /me/mfa:
    get:
      operationId: identityMFAStatus
      responses:
        "200":
          {
            description: MFA status,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/MFAStatus" } },
              },
          }
    post:
      operationId: identityBeginMFASetup
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/MFASetupRequest" } },
            },
        }
      responses:
        "200":
          {
            description: One-time TOTP setup secret,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/MFASetup" } },
              },
          }
        "401": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: identityDisableMFA
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/MFAProtectedRequest" },
                },
            },
        }
      responses:
        "204": { description: MFA disabled and all sessions revoked }
        "401": { $ref: "#/components/responses/Problem" }
  /me/mfa/confirm:
    post:
      operationId: identityConfirmMFASetup
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/MFACodeRequest" } },
            },
        }
      responses:
        "200":
          {
            description: MFA enabled and one-time recovery codes returned,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/RecoveryCodes" } },
              },
          }
        "401": { $ref: "#/components/responses/Problem" }
  /me/mfa/recovery-codes:
    post:
      operationId: identityRegenerateRecoveryCodes
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/MFAProtectedRequest" },
                },
            },
        }
      responses:
        "200":
          {
            description: Previous codes invalidated and new one-time recovery codes returned,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/RecoveryCodes" } },
              },
          }
        "401": { $ref: "#/components/responses/Problem" }
  /workspaces:
    post:
      operationId: tenancyCreateWorkspace
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/CreateWorkspaceRequest" },
                },
            },
        }
      responses:
        "201":
          {
            description: Workspace and owner membership created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/WorkspaceDetails" } },
              },
          }
        "409": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /invitations/accept:
    post:
      operationId: tenancyAcceptInvitation
      parameters: [{ $ref: "#/components/parameters/CSRFToken" }]
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/AcceptInvitationRequest" },
                },
            },
        }
      responses:
        "200":
          {
            description: Membership created or reactivated,
            content:
              {
                application/json:
                  {
                    schema: { $ref: "#/components/schemas/MembershipMutation" },
                  },
              },
          }
        "404": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/invitations:
    post:
      operationId: tenancyInviteMember
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/InviteMemberRequest" },
                },
            },
        }
      responses:
        "201":
          {
            description: Invitation created; plaintext token is returned once,
            content:
              {
                application/json:
                  {
                    schema:
                      { $ref: "#/components/schemas/WorkspaceInvitation" },
                  },
              },
          }
        "403": { $ref: "#/components/responses/Problem" }
        "422": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/members:
    get:
      operationId: tenancyListMembers
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      responses:
        "200":
          {
            description: Workspace members,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/WorkspaceMember" },
                      },
                  },
              },
          }
        "403": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/members/{membershipId}/role:
    patch:
      operationId: tenancyUpdateMemberRole
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - {
            name: membershipId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/MemberRoleRequest" } },
            },
        }
      responses:
        "200":
          {
            description: Membership role changed,
            content:
              {
                application/json:
                  {
                    schema: { $ref: "#/components/schemas/MembershipMutation" },
                  },
              },
          }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
...
  /workspaces/{workspaceId}/members/{membershipId}/status:
    patch:
      operationId: tenancyUpdateMemberStatus
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - {
            name: membershipId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema: { $ref: "#/components/schemas/MemberStatusRequest" },
                },
            },
        }
      responses:
        "200":
          {
            description: Membership status changed,
            content:
              {
                application/json:
                  {
                    schema: { $ref: "#/components/schemas/MembershipMutation" },
                  },
              },
          }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/defaults:
    patch:
      operationId: tenancyUpdateWorkspaceDefaults
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/IfMatch" }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/WorkspaceDefaultsRequest" },
                },
            },
        }
      responses:
        "200":
          {
            description: Workspace defaults changed,
            headers: { ETag: { schema: { type: string } } },
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/WorkspaceDetails" } },
              },
          }
        "412": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/me/locale:
    patch:
      operationId: tenancySetMyWorkspaceLocale
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                {
                  schema:
                    { $ref: "#/components/schemas/WorkspaceLocaleRequest" },
                },
            },
        }
      responses:
        "200":
          {
            description: Per-workspace locale override changed,
            content:
              {
                application/json:
                  {
                    schema: { $ref: "#/components/schemas/MembershipMutation" },
                  },
              },
          }
  /workspaces/{workspaceId}/departments:
    get:
      operationId: tenancyListDepartments
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      responses:
        "200":
          description: Departments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Department" }
    post:
      operationId: tenancyCreateDepartment
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateDepartmentRequest" }
      responses:
        "201":
          description: Department created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Department" }
  /workspaces/{workspaceId}/departments/{teamId}/members:
    get:
      operationId: tenancyListDepartmentMembers
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Department membership list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/TeamMember" }
  /workspaces/{workspaceId}/departments/{teamId}/members/{membershipId}:
    put:
      operationId: tenancyAddDepartmentMember
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: membershipId, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "204": { description: Member added to department }
    delete:
      operationId: tenancyRemoveDepartmentMember
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { name: teamId, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: membershipId, in: path, required: true, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "204": { description: Member removed from department }
  /workspaces/{workspaceId}/teams:
    get:
      operationId: tenancyListTeams
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      responses:
        "200":
          {
            description: Teams,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/Team" },
                      },
                  },
              },
          }
    post:
      operationId: tenancyCreateTeam
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/CSRFToken" }
      requestBody:
        {
          required: true,
          content:
            {
              application/json:
                { schema: { $ref: "#/components/schemas/CreateTeamRequest" } },
            },
        }
      responses:
        "201":
          {
            description: Team created,
            content:
              {
                application/json:
                  { schema: { $ref: "#/components/schemas/Team" } },
              },
          }
  /workspaces/{workspaceId}/teams/{teamId}/members:
    get:
      operationId: tenancyListTeamMembers
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - {
            name: teamId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
      responses:
        "200":
          {
            description: Team membership list,
            content:
              {
                application/json:
                  {
                    schema:
                      {
                        type: array,
                        items: { $ref: "#/components/schemas/TeamMember" },
                      },
                  },
              },
          }
  /workspaces/{workspaceId}/teams/{teamId}/members/{membershipId}:
    put:
      operationId: tenancyAddTeamMember
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - {
            name: teamId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: membershipId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "204": { description: Member added to team }
    delete:
      operationId: tenancyRemoveTeamMember
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - {
            name: teamId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - {
            name: membershipId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - { $ref: "#/components/parameters/CSRFToken" }
      responses:
        "204": { description: Member removed from team }
  /workspaces/{workspaceId}/dashboard:
    get:
      operationId: reportingDashboard
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters: [{ $ref: "#/components/parameters/WorkspaceId" }]
      responses:
        "200":
          description: Workspace KPI snapshot
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dashboard" }
        "404": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/contacts:
    get:
      operationId: customersListContacts
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/Cursor" }
        - {
            name: limit,
            in: query,
            schema: { type: integer, minimum: 1, maximum: 100, default: 50 },
          }
        - { name: q, in: query, schema: { type: string, maxLength: 120 } }
        - { name: status, in: query, schema: { type: string, maxLength: 40 } }
      responses:
        "200":
          description: Cursor-paginated contacts
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContactPage" }
        "404": { $ref: "#/components/responses/Problem" }
    post:
      operationId: customersCreateContact
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/IdempotencyKey" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateContact" }
      responses:
        "201":
          description: Contact created
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }
        "400": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }
  /workspaces/{workspaceId}/contacts/{contactId}:
    get:
      operationId: customersGetContact
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/ContactId" }
      responses:
        "200":
          description: Contact details
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ContactDetails" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      operationId: customersUpdateContact
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/ContactId" }
        - { $ref: "#/components/parameters/IfMatch" }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateContact" }
      responses:
        "200":
          description: Contact updated
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
    delete:
      operationId: customersDeleteContact
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/ContactId" }
        - { $ref: "#/components/parameters/IfMatch" }
      responses:
        "204": { description: Contact moved to trash }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
...
  /workspaces/{workspaceId}/companies:
    get:
      operationId: customersListCompanies
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - { $ref: "#/components/parameters/Cursor" }
        - { $ref: "#/components/parameters/Limit" }
        - { name: q, in: query, schema: { type: string, maxLength: 120 } }
        - { name: status, in: query, schema: { type: string, maxLength: 40 } }
      responses:
        "200":
          description: Cursor-paginated companies
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompanyPage" }
    post:
      operationId: customersCreateCompany
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters:
        [
          { $ref: "#/components/parameters/WorkspaceId" },
          { $ref: "#/components/parameters/IdempotencyKey" },
        ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCompany" }
      responses:
        "201":
          description: Company created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Company" }
        "400": { $ref: "#/components/responses/Problem" }
...
  /workspaces/{workspaceId}/activities/{activityId}/complete:
    patch:
      operationId: activitiesComplete
      security: [{ cookieSession: [] }, { bearerAPIKey: [] }]
      parameters:
        - { $ref: "#/components/parameters/WorkspaceId" }
        - {
            name: activityId,
            in: path,
            required: true,
            schema: { type: string, format: uuid },
          }
        - { $ref: "#/components/parameters/IfMatch" }
      responses:
        "200":
          description: Task completed
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Activity" }
        "404": { $ref: "#/components/responses/Problem" }
        "412": { $ref: "#/components/responses/Problem" }
...
  /workspaces/{workspaceId}/audit:
    get:
      operationId: auditList
      parameters:
        [
          { $ref: "#/components/parameters/WorkspaceId" },
          { $ref: "#/components/parameters/Cursor" },
        ]
      responses:
        "200":
          description: Append-only audit events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AuditEvent" }
...
components:
  securitySchemes:
    cookieSession:
      type: apiKey
      in: cookie
      name: veltrix_crm_session
      description: Browser session cookie. Production deployments use the __Host- prefix.
    bearerAPIKey:
      type: http
      scheme: bearer
      bearerFormat: crm1 workspace-scoped API key
...
  responses:
    Problem:
      description: RFC Problem Details with stable localizable code
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    LoginRequest:
      type: object
      additionalProperties: false
      required: [email, password]
      properties:
        email: { type: string, format: email, maxLength: 254 }
        password: { type: string, minLength: 8, maxLength: 1024 }
...
    SessionView:
      type: object
      required: [user, workspaces]
      properties:
        user: { $ref: "#/components/schemas/User" }
        workspaces:
          { type: array, items: { $ref: "#/components/schemas/Workspace" } }
...
    MFAChallenge:
      type: object
      additionalProperties: false
      required: [mfaRequired, challengeToken, expiresAt]
      properties:
        mfaRequired: { type: boolean, const: true }
        challengeToken: { type: string, minLength: 40, maxLength: 256 }
        expiresAt: { type: string, format: date-time }
...
    PasswordResetRequest:
      type: object
      additionalProperties: false
      required: [email]
      properties: { email: { type: string, format: email, maxLength: 254 } }
...